Avoid using `study.area.binary` as internal variable
study.area.binary is the value passed to binary.mask argument of adapt.binarymask() function. I would not use it within this function as it makes difficult to understand and prone to errors.
We can continue to use binary.mask variable instead as it is not used afterwards in the function.